src/test/java/za/co/game/PokerGameTest.java
package za.co.game;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.AfterEach;
import java.io.ByteArrayOutputStream;
import java.io.PrintStream;

import static org.junit.jupiter.api.Assertions.assertTrue;

public class PokerGameTest {


    private final ByteArrayOutputStream outContent = new ByteArrayOutputStream();
    private final PrintStream originalOut = System.out;

    @BeforeEach
    public void setUpStreams() {
        System.setOut(new PrintStream(outContent));
    }

    @AfterEach
    public void restoreStreams() {
        System.setOut(originalOut);
    }

    @Test
    public void testStartGame() {
        PokerGame game = new PokerGame();
        game.startGame();
        String output = outContent.toString();
        assertTrue(output.contains("Your hand: "));
        assertTrue(output.contains("You have: "));
    }

}
